Drop v1 service & permission interfaces; route OAuth through v2 oauth service
The v1 service proxy (/_services/, /api/services, secrets-style
provides/requires_services) and v1 permissions (/api/permissions,
service_access_rules, permissions table) are removed. Apps now use
[[services.v2.provides]] / [[services.v2.consumes]] exclusively.

The router-internal GitHub OAuth helper (used to clone/pull private
repos) is rewritten to call the v2 oauth service over HTTP loopback,
authenticating as a synthetic OPENHOST consumer with a hard-coded grant.

Migration v0008 drops the dead service_providers and permissions tables
on existing zones.
