+++ compute_space/src/compute_space/core/oauth.py
@@ -1,0 +1,73 @@
+"""Router-internal OAuth token helper.
+
+The router occasionally needs to act as an OAuth client itself — most
+notably to clone or pull private GitHub repos on behalf of the operator.
+It does this by calling the v2 oauth service (provider app
+``github.com/imbue-openhost/openhost/services/oauth``) over HTTP loopback,
+authenticating as a synthetic ``OPENHOST`` consumer with a hard-coded
+grant for the requested provider+scopes.
+"""
+
+import json
+import sqlite3
+
+import httpx
+
+from compute_space.core.services_v2 import ServiceNotAvailable
+from compute_space.core.services_v2 import resolve_provider
+from compute_space.core.util import assert_str
+from compute_space.db import get_db
+
+OAUTH_SERVICE_URL = "github.com/imbue-openhost/openhost/services/oauth"
+
+# Synthetic identity used in the X-OpenHost-Consumer-* headers when the
+# router itself is the consumer. Real apps have non-zero base58 ids.
+ROUTER_CONSUMER_ID = "0"
+ROUTER_CONSUMER_NAME = "OPENHOST"
+
+
+class OAuthAuthorizationRequired(Exception):
+    def __init__(self, authorize_url: str):
+        self.authorize_url = authorize_url
+
+
+async def get_oauth_token(
+    provider: str,
+    scopes: list[str],
+    return_to: str,
+    db: sqlite3.Connection | None = None,
+) -> str:
+    """Fetch an OAuth access token from the v2 oauth service.
+
+    Raises:
+        ServiceNotAvailable: The oauth service isn't installed/running, or didn't respond.
+        OAuthAuthorizationRequired: User authorization is needed (carries authorize_url).
+    """
+    if db is None:
+        db = get_db()
+    _, port, _, endpoint = resolve_provider(OAUTH_SERVICE_URL, ">=0", db)
+
+    grant_payload = {"provider": provider, "scopes": list(scopes)}
+    permissions_header = json.dumps([{"grant": grant_payload, "scope": "global", "provider_app_id": None}])
+    url = f"http://127.0.0.1:{port}{endpoint.rstrip('/')}/token"
+    try:
+        async with httpx.AsyncClient(timeout=5) as client:
+            resp = await client.post(
+                url,
+                json={"provider": provider, "scopes": list(scopes), "return_to": return_to},
+                headers={
+                    "Accept": "application/json",
+                    "X-OpenHost-Consumer-Id": ROUTER_CONSUMER_ID,
+                    "X-OpenHost-Consumer-Name": ROUTER_CONSUMER_NAME,
+                    "X-OpenHost-Permissions": permissions_header,
+                },
+            )
+    except httpx.HTTPError as e:
+        raise ServiceNotAvailable(f"OAuth service is not responding: {e}") from e
+
+    data = resp.json()
+    if resp.status_code == 200:
+        return assert_str(data["access_token"])
+    if resp.status_code == 401 and data.get("status") == "authorization_required":
+        raise OAuthAuthorizationRequired(data["authorize_url"])
+    raise ServiceNotAvailable(f"OAuth service returned unexpected status {resp.status_code}: {resp.text}")
